pyproject.toml
[project]
name = "omglol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
